+++ lib/tasks/install/nginx.rb
@@ -1,10 +1,11 @@
 module Rsm
   module Install
     class Nginx < Rsm::Base
-      attr_reader :domain, :auth_basic, :auth_basic_realm, :auth_basic_user_file
+      attr_reader :domain, :rewrite_www, :auth_basic, :auth_basic_realm, :auth_basic_user_file
 
       class_option :nginx_root, :default => "/etc/nginx", :aliases => "-n", :desc => "Nginx configuration root"
       class_option :domain, :aliases => "-d", :desc => "Server's domain"
+      class_option :rewrite_www, :type => :boolean, :default => false, :desc => "Added www-subdomain rewriting"
 
       class_option :auth_basic, :type => :boolean, :default => false, :aliases => "-a", :desc => "Use auth_basic"
       class_option :auth_basic_realm, :desc => "auth_basic realm or capitalized NAME unless set"
@@ -23,7 +24,9 @@ module Rsm
 
       def nginx_unicorn_config
         @domain = options[:domain]
-        @domain = `hostname` unless @domain
+        @domain = "#{application_name}.`hostname`" unless @domain
+
+        @rewrite_www = options[:rewrite_www]
 
         @auth_basic = options[:auth_basic]
         @auth_basic_realm = options[:auth_basic_realm]